refactor: Se reorganizan los hooks de categories a sus carpetas respectivas
